+++ src/asm.h
@@ -27,7 +27,7 @@ void asm_gen_variable_def(struct Asm *as, struct Node *node);
 // Add a string label to the data section.
 void asm_gen_store_string(struct Asm *as, struct Node *node);
 // Add data to stack; node must be a literal
-void asm_gen_add_to_stack(struct Asm *as, struct Node *node, size_t stack_offset);
+void asm_gen_add_to_stack(struct Asm *as, struct Node *node, int stack_offset);
 
 void asm_gen_function_call(struct Asm *as, struct Node *node);
